Add security to the profile view and add, edit and delete products views: login is required and for the product admin views you need to be superuser

+++ products/views.py
@@ -3,12 +3,13 @@ from django.contrib import messages
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Product_Category 
+from django.contrib.auth.decorators import login_required
+
 from django.conf import settings
 from .forms import ProductForm
 
 
 # Create your views here.
-
 def all_products(request):
     """ A view to show all products, including sorting and search queries """
 
@@ -63,6 +64,7 @@ def all_products(request):
 
     return render(request, 'products/products.html', context)
 
+
 def product_detail(request, product_id):
     """ A view to show product details """
     print("product_id", product_id)
@@ -77,8 +79,14 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
+
+@login_required
 def add_product(request):
     """Add a product to the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Super User permissions required')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -97,8 +105,13 @@ def add_product(request):
 
     return render(request, template, context)
 
+
+@login_required
 def edit_product(request, product_id):
     """Update a product n the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Super User permissions required')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
@@ -121,8 +134,13 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+
+@login_required
 def delete_product(request, product_id):
     """Delete a product from the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Super User permissions required')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
